requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
coverage==7.6.12
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-flex-fields==1.0.2
flake8==7.1.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isort==6.0.1
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg==3.2.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.6.3
zope.interface==7.2
